requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
idna==2.9
json-tools==0.4.1
jsonpath-rw==1.4.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
ply==3.11
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.1
pytest-html==2.1.1
pytest-metadata==1.9.0
redis==3.5.0
requests==2.23.0
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
